requirements.txt
opencv-python
numpy
pandas
tensorflow
scikit-learn
openpyxl
mysql-connector-python
sqlalchemy